Split up code into more modules
