+++ src/codegen.rs
@@ -40,93 +40,6 @@ impl<T: PartialGerberCode<W>, W: Write> PartialGerberCode<W> for Option<T> {
     }
 }
 
-impl<W: Write> GerberCode<W> for Operation {
-    fn serialize(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            Operation::Interpolate(ref coords, ref offset) => {
-                coords.serialize_partial(writer)?;
-                offset.serialize_partial(writer)?;
-                write!(writer, "D01*\n")?;
-            },
-            Operation::Move(ref coords) => {
-                coords.serialize_partial(writer)?;
-                write!(writer, "D02*\n")?;
-            },
-            Operation::Flash(ref coords) => {
-                coords.serialize_partial(writer)?;
-                write!(writer, "D03*\n")?;
-            }
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> GerberCode<W> for DCode {
-    fn serialize(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            DCode::Operation(ref operation) => operation.serialize(writer)?,
-            DCode::SelectAperture(code) => write!(writer, "D{}*\n", code)?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> GerberCode<W> for InterpolationMode {
-    fn serialize(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            InterpolationMode::Linear => write!(writer, "G01*\n")?,
-            InterpolationMode::ClockwiseCircular => write!(writer, "G02*\n")?,
-            InterpolationMode::CounterclockwiseCircular => write!(writer, "G03*\n")?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> GerberCode<W> for QuadrantMode {
-    fn serialize(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            QuadrantMode::Single => write!(writer, "G74*\n")?,
-            QuadrantMode::Multi => write!(writer, "G75*\n")?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> GerberCode<W> for GCode {
-    fn serialize(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            GCode::InterpolationMode(ref mode) => mode.serialize(writer)?,
-            GCode::RegionMode(enabled) => if enabled {
-                write!(writer, "G36*\n")?;
-            } else {
-                write!(writer, "G37*\n")?;
-            },
-            GCode::QuadrantMode(ref mode) => mode.serialize(writer)?,
-            GCode::Comment(ref comment) => write!(writer, "G04 {} *\n", comment)?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> GerberCode<W> for MCode {
-    fn serialize(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            MCode::EndOfFile => write!(writer, "M02*\n")?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for Unit {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            Unit::Millimeters => write!(writer, "MM")?,
-            Unit::Inches => write!(writer, "IN")?,
-        };
-        Ok(())
-    }
-}
-
 impl<W: Write> GerberCode<W> for Command {
     fn serialize(&self, writer: &mut W) -> GerberResult<()> {
         match *self {
@@ -188,94 +101,6 @@ impl<W: Write> GerberCode<W> for ExtendedCode {
                 write!(writer, "%TD{}*%\n", attr)?;
             },
             _ => unimplemented!(),
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for ApertureDefinition {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        write!(writer, "{}", self.code)?;
-        self.aperture.serialize_partial(writer)?;
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for Circle {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match self.hole_diameter {
-            Some(hole_diameter) => {
-                write!(writer, "{}X{}", self.diameter, hole_diameter)?;
-            },
-            None => write!(writer, "{}", self.diameter)?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for Rectangular {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match self.hole_diameter {
-            Some(hole_diameter) => write!(writer, "{}X{}X{}", self.x, self.y, hole_diameter)?,
-            None => write!(writer, "{}X{}", self.x, self.y)?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for Polygon {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match (self.rotation, self.hole_diameter) {
-            (Some(rot), Some(hd)) => write!(writer, "{}X{}X{}X{}", self.diameter, self.vertices, rot, hd)?,
-            (Some(rot), None) => write!(writer, "{}X{}X{}", self.diameter, self.vertices, rot)?,
-            (None, Some(hd)) => write!(writer, "{}X{}X0X{}", self.diameter, self.vertices, hd)?,
-            (None, None) => write!(writer, "{}X{}", self.diameter, self.vertices)?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for Aperture {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            Aperture::Circle(ref circle) => {
-                write!(writer, "C,")?;
-                circle.serialize_partial(writer)?;
-            },
-            Aperture::Rectangle(ref rectangular) => {
-                write!(writer, "R,")?;
-                rectangular.serialize_partial(writer)?;
-            },
-            Aperture::Obround(ref rectangular) => {
-                write!(writer, "O,")?;
-                rectangular.serialize_partial(writer)?;
-            },
-            Aperture::Polygon(ref polygon) => {
-                write!(writer, "P,")?;
-                polygon.serialize_partial(writer)?;
-            },
-            Aperture::Other(ref string) => write!(writer, "{}", string)?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for Polarity {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            Polarity::Clear => write!(writer, "C")?,
-            Polarity::Dark => write!(writer, "D")?,
-        };
-        Ok(())
-    }
-}
-
-impl<W: Write> PartialGerberCode<W> for StepAndRepeat {
-    fn serialize_partial(&self, writer: &mut W) -> GerberResult<()> {
-        match *self {
-            StepAndRepeat::Open { repeat_x: rx, repeat_y: ry, distance_x: dx, distance_y: dy } =>
-                write!(writer, "X{}Y{}I{}J{}", rx, ry, dx, dy)?,
-            StepAndRepeat::Close => {},
         };
         Ok(())
     }
